COM-1509 - Fix review
