+++ src/helpers/activities.js
@@ -3,6 +3,7 @@ const { ObjectID } = require('mongodb');
 const Activity = require('../models/Activity');
 const Step = require('../models/Step');
 const Card = require('../models/Card');
+const ActivityHistory = require('../models/ActivityHistory');
 
 exports.getActivity = async activityId => Activity.findOne({ _id: activityId })
   .populate({ path: 'cards', select: '-__v -createdAt -updatedAt' })
@@ -32,3 +33,13 @@ exports.addActivity = async (stepId, payload) => {
 
 exports.detachActivity = async (stepId, activityId) =>
   Step.updateOne({ _id: stepId }, { $pull: { activities: activityId } });
+
+exports.getActivityHistory = async (activityId, userId) => {
+  const activityHistories = await ActivityHistory
+    .find({ activity: activityId, user: userId })
+    .sort([['date', -1]])
+    .limit(1)
+    .lean();
+
+  return activityHistories[0];
+};
