+++ src/routes/activities.js
@@ -3,7 +3,7 @@
 const Joi = require('joi');
 Joi.objectId = require('joi-objectid')(Joi);
 const { authorizeCardAdd, authorizeActivityUpdate } = require('./preHandlers/activities');
-const { getById, update, addCard } = require('../controllers/activityController');
+const { getById, update, addCard, getActivityHistory } = require('../controllers/activityController');
 const { CARD_TEMPLATES } = require('../models/Card');
 
 exports.plugin = {
@@ -50,6 +50,18 @@ exports.plugin = {
         pre: [{ method: authorizeCardAdd }],
       },
       handler: addCard,
+    });
+
+    server.route({
+      method: 'GET',
+      path: '/{_id}/activityhistory',
+      options: {
+        validate: {
+          params: Joi.object({ _id: Joi.objectId().required() }),
+        },
+        auth: { mode: 'required' },
+      },
+      handler: getActivityHistory,
     });
   },
 };
