+++ src/routes/activityHistories.js
@@ -2,7 +2,7 @@
 
 const Joi = require('joi');
 Joi.objectId = require('joi-objectid')(Joi);
-const { addActivityHistory, getByActivityId } = require('../controllers/activityHistoryController');
+const { addActivityHistory } = require('../controllers/activityHistoryController');
 const { authorizeAddActivityHistory } = require('./preHandlers/activityHistories');
 
 exports.plugin = {
@@ -26,18 +26,6 @@ exports.plugin = {
         pre: [{ method: authorizeAddActivityHistory }],
       },
       handler: addActivityHistory,
-    });
-
-    server.route({
-      method: 'GET',
-      path: '/{_id}',
-      options: {
-        validate: {
-          params: Joi.object({ _id: Joi.objectId().required() }),
-        },
-        auth: { mode: 'required' },
-      },
-      handler: getByActivityId,
     });
   },
 };
