+++ tests/integration/activityHistories.test.js
@@ -15,95 +15,6 @@ describe('NODE ENV', () => {
   });
 });
 
-describe('ACTIVITY ROUTES - GET /activityhistories/{_id}', () => {
-  let authToken = null;
-  beforeEach(populateDB);
-  const activityId = activitiesList[0]._id;
-
-  describe('VENDOR_ADMIN', () => {
-    beforeEach(async () => {
-      authToken = await getToken('vendor_admin');
-    });
-
-    it('should get activityhistory', async () => {
-      await app.inject({
-        method: 'POST',
-        url: '/activityhistories',
-        payload: {
-          user: activityHistoriesUsersList[0],
-          activity: activityId,
-          questionnaireAnswersList: [
-            { card: cardsList[0]._id, answer: 'skusku' },
-          ],
-        },
-        headers: { 'x-access-token': authToken },
-      });
-
-      const response = await app.inject({
-        method: 'GET',
-        url: `/activityhistories/${activityId.toHexString()}`,
-        headers: { 'x-access-token': authToken },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.result.data.activityHistory).toEqual(expect.objectContaining({
-        _id: expect.any(ObjectID),
-        user: activityHistoriesUsersList[0],
-        activity: activityId,
-        date: expect.any(Date),
-        questionnaireAnswersList: [
-          { _id: expect.any(ObjectID), card: cardsList[0]._id, answer: 'skusku' },
-        ],
-      }));
-    });
-
-    it('should return a 200 even if no activityhistory', async () => {
-      const response = await app.inject({
-        method: 'GET',
-        url: `/activityhistories/${new ObjectID()}`,
-        headers: { 'x-access-token': authToken },
-      });
-
-      expect(response.statusCode).toBe(200);
-      expect(response.result.data.activityHistory).toEqual(undefined);
-    });
-  });
-
-  it('should return 401 if user is not authenticate', async () => {
-    const response = await app.inject({
-      method: 'GET',
-      url: `/activityhistories/${activityId.toHexString()}`,
-    });
-
-    expect(response.statusCode).toBe(401);
-  });
-
-  describe('Other roles', () => {
-    const roles = [
-      { name: 'helper', expectedCode: 200 },
-      { name: 'auxiliary', expectedCode: 200 },
-      { name: 'auxiliary_without_company', expectedCode: 200 },
-      { name: 'coach', expectedCode: 200 },
-      { name: 'client_admin', expectedCode: 200 },
-      { name: 'training_organisation_manager', expectedCode: 200 },
-      { name: 'trainer', expectedCode: 200 },
-    ];
-
-    roles.forEach((role) => {
-      it(`should return ${role.expectedCode} as user is ${role.name}`, async () => {
-        authToken = await getToken(role.name);
-        const response = await app.inject({
-          method: 'GET',
-          url: `/activityhistories/${activityId.toHexString()}`,
-          headers: { 'x-access-token': authToken },
-        });
-
-        expect(response.statusCode).toBe(role.expectedCode);
-      });
-    });
-  });
-});
-
 describe('ACTIVITY HISTORIES ROUTES - POST /activityhistories/', () => {
   let authToken = null;
